app/requirements.txt
pydantic==1.6.1
fastapi==0.61.1
PyJWT==1.7.1
